functions/user_search.py
# Importa o módulo discord para interagir com a API do Discord
import discord
# Importa o submódulo app_commands do discord para criar comandos de aplicação (slash commands)
from discord import app_commands
# Importa o módulo logging para registrar eventos e erros
import logging

# Configura o logger para este módulo
logger = logging.getLogger('bot')

# Função para configurar os comandos de pesquisa de usuário na árvore de comandos
def setup_user_search_commands(tree: app_commands.CommandTree):
    # Registra no log o início da configuração dos comandos
    logger.info("Configurando comandos de usuário...")
    try:
        # Aqui você pode adicionar comandos diretamente à árvore, se necessário
        # Por exemplo:
        # @tree.command(name="exemplo", description="Um comando de exemplo")
        # async def exemplo_comando(interaction: discord.Interaction):
        #     await interaction.response.send_message("Este é um comando de exemplo.")

        logger.info("Comandos de usuário configurados com sucesso")
    except Exception as e:
        # Em caso de erro, registra o erro no log
        logger.error(f"Erro ao configurar comandos de usuário: {e}", exc_info=True)

# Resumo do arquivo:
# Este arquivo define uma estrutura para comandos de pesquisa de usuário em um bot do Discord.
# Ele importa módulos necessários, configura um logger, e fornece uma função para configurar
# esses comandos na árvore de comandos do bot. O código inclui tratamento de erros e logging
# para facilitar a depuração e monitoramento do funcionamento do bot.
