requirements.txt
discord.py>=2.0.0
psycopg2-binary>=2.9.3
asyncio>=3.4.3
python-dotenv>=0.19.0